Add basic example
A basic example to show how to load power grid model into PSS/E, do the
power flow analysis, and the dynamic simulation.
This file is destined to run in Python. Slightly changes can be made to
enable it to run in PSS/E automation tools.
Need to add the ploting and data restoration functions later on.